backend/requirements.txt
Django==2.2.19
django-cors-headers
django-colorfield
djangorestframework
djangorestframework-simplejwt==4.7.2
djoser 
flake8==5.0.4
importlib-metadata==4.2.0
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
sorl-thumbnail
typing_extensions==4.4.0
zipp==3.12.1
